butfit/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
Django==4.0.1
django-cors-headers==3.10.1
mysqlclient==2.1.0
pycparser==2.21
PyJWT==2.3.0
six==1.16.0
sqlparse==0.4.2
